Refatoração de terça feira
